Update el.json
Translated the presets, plus 'clip'.

I didn't translate 'Sin City' since the effect is based on the looks of the movie and Sin City is how people mostly know it in Greece. Also left 'Jarques' as is. 

Should they be translated in Greek anyway? Is there a recommended approach on that? Jarques would just be written with Greek characters but Sin City is translatable.